refactor(publish): 抽出 PublishTab.vue，ServerControlPanel 暫為過渡外殼
